+++ src/Service/Generator/ProductGenerator.php
@@ -127,10 +127,12 @@ class ProductGenerator
      * @param string $category
      * @param string $salesChannel
      * @param int $descriptionLength
+     * @param string $variantPropertyGroupId
+     * @param string[] $imageStyles
      * @throws \JsonException
      * @return void
      */
-    public function generate(string $keywords, int $maxCount, string $category, string $salesChannel, int $descriptionLength, string $variantPropertyGroupId)
+    public function generate(string $keywords, int $maxCount, string $category, string $salesChannel, int $descriptionLength, string $variantPropertyGroupId, array $imageStyles)
     {
         if (empty($keywords)) {
             throw new \Exception('No keywords provided. Please tell the plugin what to generate.');
@@ -145,14 +147,14 @@ class ProductGenerator
         $prompt .= 'the following properties should be generated.' . PHP_EOL;
         $prompt .= 'Every resulting line should be in the order and sort provided below:' . PHP_EOL;
         $prompt .= PHP_EOL;
-        $prompt .= 'product count.' . PHP_EOL;
-        $prompt .= 'product number code. should be 24 unique random alphanumeric.' . PHP_EOL;
-        $prompt .= 'name of the product.' . PHP_EOL;
-        $prompt .= 'description (about ' . $descriptionLength . ' characters).' . PHP_EOL;
-        $prompt .= 'price value (no currency just number).' . PHP_EOL;
-        $prompt .= 'EAN code.' . PHP_EOL;
-        $prompt .= 'SEO description (max 100 characters).' . PHP_EOL;
-        $prompt .= 'variant indicator (1 if variants make sense for the product, 0 if it does not make sense).' . PHP_EOL;
+        $prompt .= '- product count.' . PHP_EOL;
+        $prompt .= '- product number code. should be 24 unique random alphanumeric (always unique, maybe consider datetime now).' . PHP_EOL;
+        $prompt .= '- name of the product.' . PHP_EOL;
+        $prompt .= '- description (about ' . $descriptionLength . ' characters).' . PHP_EOL;
+        $prompt .= '- price value (no currency just number).' . PHP_EOL;
+        $prompt .= '- EAN code.' . PHP_EOL;
+        $prompt .= '- SEO description (max 100 characters).' . PHP_EOL;
+        $prompt .= '- variant indicator (1 if variants make sense for the product, 0 if it does not make sense).' . PHP_EOL;
         $prompt .= PHP_EOL;
         $prompt .= 'Please only create exactly this number of products: ' . $maxCount . PHP_EOL;
         $prompt .= PHP_EOL;
@@ -214,7 +216,7 @@ class ProductGenerator
                         $this->callback->onProductImageGenerating();
                     }
 
-                    $tmpImageFile = $this->generateImage($name, $description);
+                    $tmpImageFile = $this->generateImage($name, $description, $imageStyles);
                 } else {
                     $tmpImageFile = '';
                 }
@@ -238,9 +240,6 @@ class ProductGenerator
                     $this->callback->onProductGenerated($number, $name, $currentCount, $maxCount);
                 }
             } catch (\Exception $ex) {
-                var_dump($ex->getMessage());
-                var_dump($ex->getTraceAsString());
-
                 if ($this->callback !== null) {
                     $this->callback->onProductGenerationFailed($ex->getMessage(), $currentCount, $maxCount);
                 }
@@ -394,12 +393,15 @@ class ProductGenerator
     /**
      * @param string $productName
      * @param string $productDescription
-     * @throws \Exception
+     * @param string[] $imageStyles
+     * @throws \JsonException
      * @return string
      */
-    private function generateImage(string $productName, string $productDescription): string
+    private function generateImage(string $productName, string $productDescription, array $imageStyles): string
     {
-        $prompt = $productName . ' ' . $productDescription;
+        $prompt = "Generate image for our online shop. It has to be really good and present the product in the best way possible." . PHP_EOL;
+        $prompt .= "Exactly stick to a style that matches one of these: " . implode(', ', $imageStyles) . PHP_EOL;
+        $prompt .= "Here are instructions about the product: " . $productName . ' ' . $productDescription . '.';
 
         $url = $this->openAI->generateImage($prompt, $this->imageSize);